refactor: block sealing is now in it's own module
also merged Makefile/scripts from main branch
