+++ node/cli/src/run_as_localchain.rs
@@ -1,4 +1,5 @@
 use crate::cli::Cli;
+use frequency_service::block_sealing::frequency_dev_sealing;
 use sc_cli::SubstrateCli;
 use std::process;
 
@@ -17,6 +18,12 @@ pub fn run_as_localchain(cli: Cli) -> sc_service::Result<(), sc_cli::Error> {
 			);
 			process::exit(1);
 		}
-		frequency_service::service::frequency_dev_sealing(config, cli.sealing).map_err(Into::into)
+		frequency_dev_sealing(
+			config,
+			cli.sealing,
+			u16::from(cli.sealing_interval),
+			cli.sealing_allow_empty_blocks,
+		)
+		.map_err(Into::into)
 	})
 }
